move the get_unique_id our code back to misc
